+++ arc.h
@@ -18,7 +18,7 @@ int arc_print(void*);
 T_ARC arc_make(int arr, float c);
 
 int arc_geta(T_ARC);
-int arc_getc(T_ARC);
+double arc_getc(T_ARC);
 
 T_ARC trouve_arc(L_ARC larc_d, int a);    //func utile dans la recherche du pcc
 void arc_setc(T_ARC, double cout);
